fix: apply ssa namespace twice
In argoproj/argo-cd#13965, upgrading to client-go 0.26 uncovered an
issue with existing namespaces not preserving labels and annotations
when converting an existing namespace to use `managedNamespaceMetadata`.

From 0.26 up, in order to preserve the existing behaviour what we need
to do is to first apply the existing namespace with SSA and its
preexisting labels and annotations. This also needs to be applied with
another manager than the default one.

After that initial apply, we proceed as normal with a ServerSideApply.
This is a one-time thing for any existing namespace; we check if we need
to perform this action by detecting if there is a preexisting SSA
annotation on the namespace.

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>